+++ mlpf/model/hept.py
@@ -1,0 +1,443 @@
+import math
+import random
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+from einops import rearrange
+
+
+def trunc_normal_(tensor, mean=0.0, std=1.0, a=-2.0, b=2.0):
+    # From https://github.com/rwightman/pytorch-image-models/blob/
+    #        18ec173f95aa220af753358bf860b16b6691edb2/timm/layers/weight_init.py#L8
+    r"""Fills the input Tensor with values drawn from a truncated
+    normal distribution. The values are effectively drawn from the
+    normal distribution :math:`\mathcal{N}(\text{mean}, \text{std}^2)`
+    with values outside :math:`[a, b]` redrawn until they are within
+    the bounds. The method used for generating the random values works
+    best when :math:`a \leq \text{mean} \leq b`.
+    Args:
+        tensor: an n-dimensional `torch.Tensor`
+        mean: the mean of the normal distribution
+        std: the standard deviation of the normal distribution
+        a: the minimum cutoff value
+        b: the maximum cutoff value
+    Examples:
+        >>> w = torch.empty(3, 5)
+        >>> nn.init.trunc_normal_(w)
+    """
+
+    def norm_cdf(x):
+        # Computes standard normal cumulative distribution function
+        return (1.0 + math.erf(x / math.sqrt(2.0))) / 2.0
+
+    with torch.no_grad():
+        # Values are generated by using a truncated uniform distribution and
+        # then using the inverse CDF for the normal distribution.
+        # Get upper and lower cdf values
+        lo = norm_cdf((a - mean) / std)
+        up = norm_cdf((b - mean) / std)
+
+        # Uniformly fill tensor with values from [l, u], then translate to
+        # [2lo-1, 2up-1].
+        tensor.uniform_(2 * lo - 1, 2 * up - 1)
+
+        # Use inverse cdf transform for normal distribution to get truncated
+        # standard normal
+        tensor.erfinv_()
+
+        # Transform to proper mean, std
+        tensor.mul_(std * math.sqrt(2.0))
+        tensor.add_(mean)
+
+        # Clamp to ensure it's in the proper range
+        tensor.clamp_(min=a, max=b)
+        return tensor
+
+
+def uniform(a, b, shape, device="cpu"):
+    return (b - a) * torch.rand(shape, device=device) + a
+
+
+class E2LSH(nn.Module):
+    def __init__(self, n_hashes, n_heads, dim, r=1):
+        super(E2LSH, self).__init__()
+        self.alpha = nn.Parameter(torch.normal(0, 1, (n_heads, dim, n_hashes)))
+        self.beta = nn.Parameter(uniform(0, r, shape=(1, n_hashes)))
+        self.alpha.requires_grad = False
+        self.beta.requires_grad = False
+
+    def forward(self, vecs):
+        projection = torch.bmm(vecs, self.alpha)
+        return projection.permute(2, 0, 1)
+
+
+def lsh_mapping(e2lsh, queries, keys):
+    queries_hashed = e2lsh(queries)
+    keys_hashed = e2lsh(keys)
+    max_hash_shift = torch.max(queries_hashed.max(-1, keepdim=True).values, keys_hashed.max(-1, keepdim=True).values)
+    min_hash_shift = torch.min(queries_hashed.min(-1, keepdim=True).values, keys_hashed.min(-1, keepdim=True).values)
+    hash_shift = max_hash_shift - min_hash_shift
+    return queries_hashed, keys_hashed, hash_shift
+
+
+def batched_index_select(values: torch.Tensor, indices: torch.Tensor) -> torch.Tensor:
+    last_dim = values.shape[-1]
+    indices_expanded = rearrange(indices, "... -> ... 1").expand(*indices.shape, last_dim)
+    return values.expand(*indices_expanded.shape[:-2], *values.shape[-2:]).gather(-2, indices_expanded)
+
+
+def invert_permutation(perm: torch.Tensor) -> torch.Tensor:
+    arange = torch.arange(perm.shape[-1], device=perm.device).expand_as(perm)
+    return torch.empty_like(perm).scatter_(-1, perm, arange)
+
+
+def sort_to_buckets(x, perm, bucketsz):
+    return rearrange(
+        batched_index_select(rearrange(x, "b s d -> 1 b s d"), perm),
+        "h b (nbuckets bucketsz) d -> h b nbuckets bucketsz d",
+        bucketsz=bucketsz,
+    )
+
+
+def unsort_from_buckets(s_x, perm_inverse):
+    b_x = rearrange(s_x, "h b nbuckets bucketsz d -> h b (nbuckets bucketsz) d")
+    return batched_index_select(b_x, perm_inverse)
+
+
+def qkv_res(s_query, s_key, s_value, s_mask_q=None, s_mask_k=None, return_attn=False):
+    # Upcast to float32 for stable RBF distance and weighted sum
+    s_query, s_key, s_value = s_query.float(), s_key.float(), s_value.float()
+    q_sq_05 = -0.5 * (s_query**2).sum(dim=-1, keepdim=True)
+    k_sq_05 = -0.5 * (s_key**2).sum(dim=-1, keepdim=True)
+    clustered_dists = torch.einsum("...id,...jd->...ij", s_query, s_key)
+    clustered_dists = clustered_dists + q_sq_05 + k_sq_05.transpose(-1, -2)
+    if s_mask_k is not None:
+        clustered_dists = clustered_dists.masked_fill(s_mask_k.transpose(-1, -2) == 0, float("-inf"))
+    clustered_dists = clustered_dists.clamp(max=0.0).exp()
+    if s_mask_q is not None:
+        clustered_dists = clustered_dists * s_mask_q.float()
+    denom = clustered_dists.sum(dim=-1, keepdim=True) + (1e-20)
+    qk = clustered_dists
+    so = torch.einsum("...ij,...jd->...id", qk, s_value)
+    if return_attn:
+        return denom, so, qk
+    return denom, so
+
+
+def prep_qk(query, key, w, coords):
+    # Ensure coords are float32 for consistent precision
+    coords = coords.float()
+    qw = w.sum(dim=1).clamp(max=50).exp().sum(dim=-1)
+    new_qw_expand_dim = torch.cat([qw[:, :1], qw], dim=-1)
+
+    sqrt_w_r = torch.sqrt(2 * new_qw_expand_dim)[None] * coords[:, None]
+
+    # Combine in float32
+    q_hat = torch.cat([query.float(), sqrt_w_r], dim=-1)
+    k_hat = torch.cat([key.float(), sqrt_w_r], dim=-1)
+    return q_hat, k_hat
+
+
+def pad_to_multiple(tensor, multiple, dims=-1, value=0):
+    if not isinstance(dims, (list, tuple)):
+        dims = [dims]
+    dims = [d if d >= 0 else tensor.ndim + d for d in dims]
+    padding = [0] * (2 * tensor.ndim)
+    for d in dims:
+        size = tensor.size(d)
+        m = size // multiple
+        remainder = size - m * multiple
+        if remainder != 0:
+            padding[2 * (tensor.ndim - d - 1) + 1] = multiple - remainder
+    if all(p == 0 for p in padding):
+        return tensor
+    return F.pad(tensor, tuple(padding), value=value)
+
+
+def quantile_partition(sorted_indices, num_regions):
+    total_elements = sorted_indices.shape[-1]
+    region_size = torch.ceil(torch.tensor(total_elements, device=sorted_indices.device) / torch.as_tensor(num_regions, device=sorted_indices.device))
+    inverse_indices = torch.argsort(sorted_indices, dim=-1)
+    base = torch.arange(total_elements, device=sorted_indices.device)[None]
+    region_indices = base // region_size + 1
+    reassigned_regions = region_indices.gather(-1, inverse_indices.expand(region_indices.shape[0], -1))
+    return reassigned_regions
+
+
+def get_regions(num_regions, num_or_hashes, num_heads, num_and_hashes=2):
+    lb = 2
+    ub = 2 * num_regions ** (1 / num_and_hashes) - lb
+    regions = []
+    for _ in range(num_or_hashes * num_heads):
+        region = []
+        for _ in range(num_and_hashes):
+            a = random.uniform(lb, ub)
+            region.append(a)
+        regions.append(region)
+    regions = torch.tensor(regions)
+    regions = (num_regions / regions.prod(dim=1, keepdim=True)) ** (1 / num_and_hashes) * regions
+    regions = torch.round(regions * 3) / 3
+    return rearrange(regions, "(h c) a -> c a h", h=num_heads)
+
+
+@torch.no_grad()
+def get_geo_shift(regions_h, hash_shift, region_indices, num_or_hashes):
+    region_indices_eta, region_indices_phi = region_indices
+    q_hash_shift_eta = region_indices_eta * hash_shift
+    k_hash_shift_eta = region_indices_eta * hash_shift
+    q_hash_shift_phi = region_indices_phi * hash_shift * (torch.ceil(regions_h[0][:, None]) + 1)
+    k_hash_shift_phi = region_indices_phi * hash_shift * (torch.ceil(regions_h[0][:, None]) + 1)
+    res = torch.stack([q_hash_shift_phi + q_hash_shift_eta, k_hash_shift_phi + k_hash_shift_eta], dim=0)
+    return rearrange(res, "a (c h) n -> a c h n", c=num_or_hashes)
+
+
+class HEPTAttention(nn.Module):
+    def __init__(self, hash_dim, **kwargs):
+        super().__init__()
+        self.dim_per_head = kwargs["h_dim"]
+        self.num_heads = kwargs["num_heads"]
+        self.out_linear = nn.Linear(self.num_heads * self.dim_per_head, self.num_heads * self.dim_per_head)
+        self.block_size = kwargs.get("block_size", 100)
+        self.n_hashes = kwargs.get("n_hashes", 3)
+        self.num_w_per_dist = kwargs.get("num_w_per_dist", 10)
+        self.e2lsh = E2LSH(n_hashes=self.n_hashes, n_heads=self.num_heads, dim=hash_dim)
+
+    def forward(self, query, key, value, return_attn=False, **kwargs):
+        # query, key, value: [N_padded, num_heads * dim_per_head]
+        query = query.view(-1, self.num_heads, self.dim_per_head)
+        key = key.view(-1, self.num_heads, self.dim_per_head)
+        value = value.view(-1, self.num_heads, self.dim_per_head)
+
+        w = rearrange(
+            kwargs["w_rpe"].weight,
+            "(h d) (r k) -> h d r k",
+            h=self.num_heads,
+            d=self.dim_per_head,
+            k=self.num_w_per_dist,
+        )
+        q_hat, k_hat = prep_qk(query, key, w, kwargs["coords"])
+
+        q_hat = rearrange(q_hat, "n h d -> h n d")
+        k_hat = rearrange(k_hat, "n h d -> h n d")
+        value = rearrange(value, "n h d -> h n d")
+
+        raw_size = kwargs["raw_size"]
+        padding_mask = kwargs.get("padding_mask", None)
+
+        q_hashed, k_hashed, hash_shift = lsh_mapping(self.e2lsh, q_hat, k_hat)
+        hash_shift = rearrange(hash_shift, "c h d -> (c h) d")
+
+        if padding_mask is not None:
+            pm_expanded = padding_mask.view(1, -1, 1)
+            q_hat = q_hat.masked_fill(pm_expanded == 0, 0.0)
+            k_hat = k_hat.masked_fill(pm_expanded == 0, 0.0)
+            value = value.masked_fill(pm_expanded == 0, 0.0)
+            q_hashed = q_hashed.masked_fill(padding_mask.view(1, 1, -1) == 0, float("inf"))
+            k_hashed = k_hashed.masked_fill(padding_mask.view(1, 1, -1) == 0, float("inf"))
+        else:
+            q_hat[:, raw_size:] = 0.0
+            k_hat[:, raw_size:] = 0.0
+            value[:, raw_size:] = 0.0
+            q_hashed[..., raw_size:] = float("inf")
+            k_hashed[..., raw_size:] = float("inf")
+
+        q_shifts, k_shifts = get_geo_shift(kwargs["regions_h"], hash_shift, kwargs["region_indices"], self.n_hashes)
+        q_hashed = q_hashed + q_shifts
+        k_hashed = k_hashed + k_shifts
+
+        q_positions = q_hashed.argsort(dim=-1)
+        k_positions = k_hashed.argsort(dim=-1)
+
+        s_query = sort_to_buckets(q_hat, q_positions, self.block_size)
+        s_key = sort_to_buckets(k_hat, k_positions, self.block_size)
+        s_value = sort_to_buckets(value, k_positions, self.block_size)
+
+        s_mask_q = None
+        s_mask_k = None
+        if padding_mask is not None:
+            pm_exp = padding_mask.view(1, -1, 1).expand(self.num_heads, -1, 1)
+            s_mask_q = sort_to_buckets(pm_exp, q_positions, self.block_size)
+            s_mask_k = sort_to_buckets(pm_exp, k_positions, self.block_size)
+        else:
+            pm = torch.ones(q_hat.shape[1], device=q_hat.device)
+            pm[raw_size:] = 0.0
+            pm_exp = pm.view(1, -1, 1).expand(self.num_heads, -1, 1)
+            s_mask_q = sort_to_buckets(pm_exp, q_positions, self.block_size)
+            s_mask_k = sort_to_buckets(pm_exp, k_positions, self.block_size)
+
+        if return_attn:
+            denom, so, qk = qkv_res(s_query, s_key, s_value, s_mask_q=s_mask_q, s_mask_k=s_mask_k, return_attn=True)
+        else:
+            denom, so = qkv_res(s_query, s_key, s_value, s_mask_q=s_mask_q, s_mask_k=s_mask_k)
+
+        q_rev_positions = invert_permutation(q_positions)
+        o = unsort_from_buckets(so, q_rev_positions)
+        logits = unsort_from_buckets(denom, q_rev_positions)
+
+        # Crop back to raw_size to match queries/keys and ensure gradients are safe
+        o = o[:, :, :raw_size]
+        logits = logits[:, :, :raw_size]
+
+        out = o.sum(dim=0) / (logits.sum(dim=0) + 1e-20)
+        out = self.out_linear(rearrange(out, "h n d -> n (h d)"))
+        if return_attn:
+            return out, (qk, q_positions, k_positions)
+        return out
+
+
+class PositionalEncoding(nn.Module):
+    def __init__(self, embedding_dim):
+        super(PositionalEncoding, self).__init__()
+        self.proj = nn.Linear(2, embedding_dim)
+
+    def forward(self, x, X):
+        # Extract eta and phi
+        eta = X[..., 2:3]
+        phi = torch.atan2(X[..., 3:4], X[..., 4:5])
+        coords = torch.cat([eta, phi], dim=-1)  # [B, N, 2]
+        return x + self.proj(coords)
+
+
+class HEPTLayer(nn.Module):
+    def __init__(self, name=None, embedding_dim=128, num_heads=16, width=512, dropout=0.1, pos=False, **kwargs):
+        super(HEPTLayer, self).__init__()
+        self.name = name
+        self.embedding_dim = embedding_dim
+        self.num_heads = num_heads
+        self.width = width
+        self.dropout_p = dropout
+        self.pos = pos
+        self.dim_per_head = embedding_dim // num_heads
+
+        self.block_size = kwargs.get("block_size", 100)
+        self.n_hashes = kwargs.get("n_hashes", 3)
+        self.num_regions = kwargs.get("num_regions", 140)
+        self.num_w_per_dist = kwargs.get("num_w_per_dist", 10)
+
+        # coords_dim = 2 (eta, phi)
+        self.coords_dim = 2
+        self.attn = HEPTAttention(self.dim_per_head + self.coords_dim, h_dim=self.dim_per_head, num_heads=num_heads, **kwargs)
+
+        self.w_q = nn.Linear(embedding_dim, embedding_dim, bias=False)
+        self.w_k = nn.Linear(embedding_dim, embedding_dim, bias=False)
+        self.w_v = nn.Linear(embedding_dim, embedding_dim, bias=False)
+
+        self.w_rpe = nn.Linear(self.num_w_per_dist * (self.coords_dim - 1), self.num_heads * self.dim_per_head)
+
+        self.regions = nn.Parameter(
+            get_regions(self.num_regions, self.n_hashes, self.num_heads),
+            requires_grad=False,
+        )
+
+        self.norm0 = nn.LayerNorm(embedding_dim)
+        self.norm1 = nn.LayerNorm(embedding_dim)
+        self.seq = nn.Sequential(nn.Linear(embedding_dim, width), nn.ELU(), nn.Linear(width, embedding_dim), nn.ELU())
+        self.dropout = nn.Dropout(dropout)
+        self.pos_embed = PositionalEncoding(embedding_dim) if pos else None
+
+    def extra_repr(self):
+        return (
+            f"embedding_dim={self.embedding_dim}, num_heads={self.num_heads}, width={self.width}, "
+            f"dropout={self.dropout_p}, pos={self.pos}, block_size={self.block_size}, "
+            f"n_hashes={self.n_hashes}, num_regions={self.num_regions}, num_w_per_dist={self.num_w_per_dist}"
+        )
+
+    def forward(self, x, mask, X_features, return_attn=False):
+        if self.pos_embed:
+            x = self.pos_embed(x, X_features)
+
+        if mask is not None:
+            mask_ = mask.unsqueeze(-1)
+
+        B, S, D = x.shape
+        device = x.device
+
+        if mask is not None:
+            x = x * mask_
+
+        # Extract eta and phi for hashing, explicitly in float32 for safety
+        eta = X_features[..., 2:3].float()
+        phi = torch.atan2(X_features[..., 3:4].float(), X_features[..., 4:5].float())
+        coords = torch.cat([eta, phi], dim=-1)  # [B, S, 2]
+
+        # Flatten batch to support HEPT which expects [N_total, D]
+        # Use large float32 offsets to ensure sample isolation
+        offsets = torch.arange(B, device=device).view(B, 1, 1).float() * 100.0
+        coords = coords + offsets
+
+        x_flat = x.view(B * S, D)
+        coords_flat = coords.view(B * S, 2)
+
+        raw_size = B * S
+        x_flat_padded = pad_to_multiple(x_flat, self.block_size, dims=0)
+        coords_flat_padded = pad_to_multiple(coords_flat, self.block_size, dims=0, value=float("inf"))
+
+        # Precompute regions and indices
+        with torch.no_grad():
+            coords_for_sort = coords_flat_padded.clone()
+            if mask is not None:
+                mask_flat = mask.view(-1)
+                mask_flat_padded = pad_to_multiple(mask_flat, self.block_size, dims=0, value=0.0)
+                coords_for_sort[mask_flat_padded == 0] = float("inf")
+            else:
+                mask_flat_padded = torch.ones_like(coords_flat_padded[..., 0])
+                mask_flat_padded[raw_size:] = 0.0
+
+            sorted_eta_idx = torch.argsort(coords_for_sort[..., 0], dim=-1)
+            sorted_phi_idx = torch.argsort(coords_for_sort[..., 1], dim=-1)
+            regions_h = rearrange(self.regions, "c a h -> a (c h)")
+            region_indices_eta = quantile_partition(sorted_eta_idx, regions_h[0][:, None])
+            region_indices_phi = quantile_partition(sorted_phi_idx, regions_h[1][:, None])
+            region_indices = [region_indices_eta, region_indices_phi]
+
+            coords_flat_padded[mask_flat_padded == 0] = 0.0
+
+        # Run attention
+        x_norm = self.norm0(x_flat_padded)
+        q, k, v = self.w_q(x_norm), self.w_k(x_norm), self.w_v(x_norm)
+
+        if return_attn:
+            attn_out, attn_data = self.attn(
+                q,
+                k,
+                v,
+                coords=coords_flat_padded,
+                w_rpe=self.w_rpe,
+                regions_h=regions_h,
+                region_indices=region_indices,
+                raw_size=raw_size,
+                padding_mask=mask_flat_padded,
+                return_attn=True,
+            )
+        else:
+            attn_out = self.attn(
+                q,
+                k,
+                v,
+                coords=coords_flat_padded,
+                w_rpe=self.w_rpe,
+                regions_h=regions_h,
+                region_indices=region_indices,
+                raw_size=raw_size,
+                padding_mask=mask_flat_padded,
+            )
+
+        # Reshape to [B, S, D]
+        attn_out = attn_out.view(B, S, D)
+
+        # Residual and FFN
+        x = x + self.dropout(attn_out)
+
+        residual = x
+        x_norm = self.norm1(x)
+        ffn_out = self.seq(x_norm)
+        ffn_out = self.dropout(ffn_out)
+
+        x = residual + ffn_out
+
+        if mask is not None:
+            x = x * mask_
+
+        if return_attn:
+            return x, attn_data
+        return x
