+++ mlpf/model/mlpf.py
@@ -1,4 +1,3 @@
-import math
 import time
 import numpy as np
 from typing import Union, List
@@ -18,6 +17,8 @@ if litept_path not in sys.path:
     sys.path.append(litept_path)
 from litept.model import LitePT
 
+from mlpf.model.hept import HEPTLayer, trunc_normal_
+
 from mlpf.conf import (
     MLPFConfig,
     Activation,
@@ -27,54 +28,6 @@ from mlpf.conf import (
     LearnedRepresentationMode,
     RegressionMode,
 )
-
-
-def trunc_normal_(tensor, mean=0.0, std=1.0, a=-2.0, b=2.0):
-    # From https://github.com/rwightman/pytorch-image-models/blob/
-    #        18ec173f95aa220af753358bf860b16b6691edb2/timm/layers/weight_init.py#L8
-    r"""Fills the input Tensor with values drawn from a truncated
-    normal distribution. The values are effectively drawn from the
-    normal distribution :math:`\mathcal{N}(\text{mean}, \text{std}^2)`
-    with values outside :math:`[a, b]` redrawn until they are within
-    the bounds. The method used for generating the random values works
-    best when :math:`a \leq \text{mean} \leq b`.
-    Args:
-        tensor: an n-dimensional `torch.Tensor`
-        mean: the mean of the normal distribution
-        std: the standard deviation of the normal distribution
-        a: the minimum cutoff value
-        b: the maximum cutoff value
-    Examples:
-        >>> w = torch.empty(3, 5)
-        >>> nn.init.trunc_normal_(w)
-    """
-
-    def norm_cdf(x):
-        # Computes standard normal cumulative distribution function
-        return (1.0 + math.erf(x / math.sqrt(2.0))) / 2.0
-
-    with torch.no_grad():
-        # Values are generated by using a truncated uniform distribution and
-        # then using the inverse CDF for the normal distribution.
-        # Get upper and lower cdf values
-        lo = norm_cdf((a - mean) / std)
-        up = norm_cdf((b - mean) / std)
-
-        # Uniformly fill tensor with values from [l, u], then translate to
-        # [2lo-1, 2up-1].
-        tensor.uniform_(2 * lo - 1, 2 * up - 1)
-
-        # Use inverse cdf transform for normal distribution to get truncated
-        # standard normal
-        tensor.erfinv_()
-
-        # Transform to proper mean, std
-        tensor.mul_(std * math.sqrt(2.0))
-        tensor.add_(mean)
-
-        # Clamp to ensure it's in the proper range
-        tensor.clamp_(min=a, max=b)
-        return tensor
 
 
 def get_activation(activation: Activation):
@@ -499,7 +452,7 @@ class MLPF(nn.Module):
 
         # Ensure sub_config is initialized if it's None (can happen if not in spec or args)
         if sub_config is None:
-            from mlpf.conf import AttentionConfig, GNNLSHConfig, LitePTConfig
+            from mlpf.conf import AttentionConfig, GNNLSHConfig, LitePTConfig, HEPTConfig
 
             if self.conv_type == ModelType.ATTENTION:
                 sub_config = AttentionConfig()
@@ -507,6 +460,8 @@ class MLPF(nn.Module):
                 sub_config = GNNLSHConfig()
             elif self.conv_type == ModelType.LITEPT:
                 sub_config = LitePTConfig()
+            elif self.conv_type == ModelType.HEPT:
+                sub_config = HEPTConfig()
             setattr(self.config, self.conv_type.value, sub_config)
 
         # Extract parameters from the sub-config
@@ -555,6 +510,10 @@ class MLPF(nn.Module):
             self.use_pre_layernorm = False
         elif self.conv_type == ModelType.LITEPT:
             self.use_pre_layernorm = False
+        elif self.conv_type == ModelType.HEPT:
+            self.use_pre_layernorm = False
+            num_heads = sub_config.num_heads
+            pos = sub_config.pos
 
         _logger.info(f"MLPF __init__ conv_type={self.conv_type} num_convs={self.num_convs} input_encoding={self.input_encoding}")
 
@@ -660,6 +619,39 @@ class MLPF(nn.Module):
                     _logger.info(f"Initializing LitePT layer {i}")
                     self.conv_id.append(LitePTLayer(name=f"litept_id_{i}", litept_config=litept_conf, embedding_dim=embedding_dim))
                     self.conv_reg.append(LitePTLayer(name=f"litept_reg_{i}", litept_config=litept_conf, embedding_dim=embedding_dim))
+            elif self.conv_type == ModelType.HEPT:
+                _logger.info("Initializing HEPT convolution layers")
+                self.conv_id = nn.ModuleList()
+                self.conv_reg = nn.ModuleList()
+                hept_conf = self.config.hept.model_dump()
+                # Remove keys that HEPTLayer constructor handles via explicit arguments
+                for key in ["num_convs", "conv_type", "embedding_dim", "width", "activation", "dropout_ff", "num_heads", "pos"]:
+                    if key in hept_conf:
+                        hept_conf.pop(key)
+                for i in range(self.num_convs):
+                    _logger.info(f"Initializing HEPT layer {i}")
+                    self.conv_id.append(
+                        HEPTLayer(
+                            name=f"hept_id_{i}",
+                            embedding_dim=embedding_dim,
+                            num_heads=num_heads,
+                            width=width,
+                            dropout=dropout_ff,
+                            pos=pos,
+                            **hept_conf,
+                        )
+                    )
+                    self.conv_reg.append(
+                        HEPTLayer(
+                            name=f"hept_reg_{i}",
+                            embedding_dim=embedding_dim,
+                            num_heads=num_heads,
+                            width=width,
+                            dropout=dropout_ff,
+                            pos=pos,
+                            **hept_conf,
+                        )
+                    )
             _logger.info("Convolution layers initialization took {:.2f}s".format(time.time() - t0))
             _logger.info("conv_id parameters: {}".format(count_parameters(self.conv_id)))
             _logger.info("conv_reg parameters: {}".format(count_parameters(self.conv_reg)))
@@ -735,7 +727,7 @@ class MLPF(nn.Module):
         if self.num_convs != 0:
             for num, conv in enumerate(self.conv_id):
                 conv_input = embedding_id if num == 0 else embeddings_id[-1]
-                if self.conv_type == ModelType.LITEPT:
+                if self.conv_type == ModelType.LITEPT or self.conv_type == ModelType.HEPT:
                     out_padded = conv(conv_input, mask, X_features)
                 else:
                     out_padded = conv(conv_input, mask, embedding_id)
@@ -743,7 +735,7 @@ class MLPF(nn.Module):
 
             for num, conv in enumerate(self.conv_reg):
                 conv_input = embedding_reg if num == 0 else embeddings_reg[-1]
-                if self.conv_type == ModelType.LITEPT:
+                if self.conv_type == ModelType.LITEPT or self.conv_type == ModelType.HEPT:
                     out_padded = conv(conv_input, mask, X_features)
                 else:
                     out_padded = conv(conv_input, mask, embedding_reg)
