+++ scripts/tallinn/train.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 #SBATCH --partition gpu
-#SBATCH --gres gpu:a100:1
-#SBATCH --mem 80G
-#SBATCH --cpus-per-task 8
+#SBATCH --gres gpu:l40:1
+#SBATCH --mem-per-gpu 80G
+#SBATCH --cpus-per-gpu 4
 #SBATCH -o logs/slurm-%x-%j-%N.out
+#SBATCH --job-name=train-cld-hits
+
 set -e
 export PF_SITE=tallinn
 
@@ -22,4 +24,4 @@ DATA_DIR=$(pixi run python3 scripts/get_param.py particleflow_spec.yaml producti
     train \
     --gpus 1 \
     --num_workers 4 \
-    --prefetch_factor 2 --model.type litept --gpu_batch_multiplier 64
+    --prefetch_factor 2 --model.type litept
